Move signature verification logic to Sginature module
Removed secure_compare specs as it is not part of the public API anymore.
HMAC specs should already cover the functionality of secure_compare.
