+++ lib/jwt.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 require 'base64'
-require 'openssl'
 require 'jwt/decode'
 require 'jwt/default_options'
 require 'jwt/encode'
@@ -17,20 +16,6 @@ module JWT
   include JWT::DefaultOptions
 
   module_function
-
-  def verify_rsa(algorithm, public_key, signing_input, signature)
-    public_key.verify(OpenSSL::Digest.new(algorithm.sub('RS', 'sha')), signature, signing_input)
-  end
-
-  def verify_ecdsa(algorithm, public_key, signing_input, signature)
-    key_algorithm = Signature::NAMED_CURVES[public_key.group.curve_name]
-    if algorithm != key_algorithm
-      raise IncorrectAlgorithm, "payload algorithm is #{algorithm} but #{key_algorithm} verification key was provided"
-    end
-
-    digest = OpenSSL::Digest.new(algorithm.sub('ES', 'sha'))
-    public_key.dsa_verify_asn1(digest.digest(signing_input), raw_to_asn1(signature, public_key))
-  end
 
   def decoded_segments(jwt, key = nil, verify = true, custom_options = {}, &keyfinder)
     raise(JWT::DecodeError, 'Nil JSON web token') unless jwt
@@ -66,7 +51,7 @@ module JWT
     raise(JWT::IncorrectAlgorithm, 'An algorithm must be specified') unless options[:algorithm]
     raise(JWT::IncorrectAlgorithm, 'Expected a different algorithm') unless algo == options[:algorithm]
 
-    verify_signature(algo, key, signing_input, signature)
+    Signature.verify(algo, key, signing_input, signature)
   end
 
   def signature_algorithm_and_key(header, payload, key, &keyfinder)
@@ -79,43 +64,6 @@ module JWT
       raise JWT::DecodeError, 'No verification key available' unless key
     end
     [header['alg'], key]
-  end
-
-  def verify_signature(algo, key, signing_input, signature)
-    verify_signature_algo(algo, key, signing_input, signature)
-  rescue OpenSSL::PKey::PKeyError
-    raise JWT::VerificationError, 'Signature verification raised'
-  ensure
-    OpenSSL.errors.clear
-  end
-
-  def verify_signature_algo(algo, key, signing_input, signature)
-    if %w(HS256 HS384 HS512).include?(algo)
-      raise(JWT::VerificationError, 'Signature verification raised') unless secure_compare(signature, Signature.sign(algo, signing_input, key))
-    elsif %w(RS256 RS384 RS512).include?(algo)
-      raise(JWT::VerificationError, 'Signature verification raised') unless verify_rsa(algo, key, signing_input, signature)
-    elsif %w(ES256 ES384 ES512).include?(algo)
-      raise(JWT::VerificationError, 'Signature verification raised') unless verify_ecdsa(algo, key, signing_input, signature)
-    else
-      raise JWT::VerificationError, 'Algorithm not supported'
-    end
-  end
-
-  # From devise
-  # constant-time comparison algorithm to prevent timing attacks
-  def secure_compare(a, b)
-    return false if a.nil? || b.nil? || a.empty? || b.empty? || a.bytesize != b.bytesize
-    l = a.unpack "C#{a.bytesize}"
-    res = 0
-    b.each_byte { |byte| res |= byte ^ l.shift }
-    res.zero?
-  end
-
-  def raw_to_asn1(signature, private_key)
-    byte_size = (private_key.group.degree + 7) / 8
-    r = signature[0..(byte_size - 1)]
-    s = signature[byte_size..-1] || ''
-    OpenSSL::ASN1::Sequence.new([r, s].map { |int| OpenSSL::ASN1::Integer.new(OpenSSL::BN.new(int, 2)) }).to_der
   end
 
   def base64url_decode(str)
